modules/partdisks: Added "support"(workaround) for the "root" seed.
FIXME: This workaround is to get the root overriden via
a seed. This is extremly useful for those who need to install the
distribution to a partition not recognized by the installer, such
as a crypted partition.

USAGE:
  - "root" seed to the path of the device
  - "root_override" seed to True

This *may* break some partdisks functions, so we are currently
using the "root_override" seed to see if this is something wanted
by the user or not.
Usage of the CLI frontend of partdisks should not be affected.

NOTE: The target root SHOULD BE ALREADY FORMATTED, and no other
root should be selected. Also, doing a refresh *may* break out
things, so the module may need a restart.