Sorta-of fixed issues when deleting all partitions in automatic mode.
A new 'clearNew' action replaces the old clear. This action destroys everything *before* creating (in memory) the partitions.
This is then more destructive than the old 'clear', but a nice warning dialog now asks the user if he/she really wants to continue.
Also refactored the partdisks library's commit() method: as the commit action is only in disks, it's pointless redo the commit
action for every partition.