modules/userhost/inst: cli: ensure we actually start the progressbar, otherwise we will enter in a crash loop...
The missing progress.start() call after the progressbar creation made the progressbar crash
when update() is called.
This commit also ensures that the new progressbar created to replace the crashed one is actually
started, otherwise we will end in a crash loop.
Thanks to Filip Danilovic for the bug report.