modules/partdisks: front/glade: on manual_frame_creator() prefer device objects in self.changed.
If the device is already in self.changed, we should take the objects directly from it, otherwise
bad things will happen when committing things (e.g. new partitions will not be physically created).
I fighted with this for months, and in the end it's only a three line fix. Wow.