modules: added 'finish' module.
This module does those little things that are required to ensure that the
target system works correctly, but that don't warrant an entire module for
themselves.

Currently the only job that this module does is to update the initramfs
image.
This fixes the boot-up issues found on Semplice 7 on freshly installed systems.