config/semplice-fullscreen: do not execute the update module.
The user hasn't access to the network-manager configuration when running from the
linstaller session, which is the main user for the semplice-fullscreen configuration.

This means that the update process works only if the configuration has been guessed
automatically (e.g. DHCP-configured LAN via ethernet) or the user has made it work manually
via tty.

Additionally, linstaller in numb has got a decent maturity and it's pretty much stable (HA!) so
we can drop the update module for this special use case.