partdisks: front/glade: force the given device and disk if on_lvm == True. (LB #4)
The change introduced in 77371d5fcd043fd7685f486795027796105519d5 made the installer fail to accept
LVM(+LUKS) configuration, as the proxy device/disk objects may not be stored in self.changed, even
if the key exists.

This commit ensures that self.changed[device.path] is rewritten using the given device and disk objects.