semplice-fullscreen: remove mirrorselect feature.
In the Installation section we can't be sure if the connection is working or
not.

While the module itself has a connection check, it's preferable not to ask for
the mirror selection at all, as the end user doesn't have any mean to check
if the internet connection is properly configured.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>